% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{debug_futurenow}
\alias{debug_futurenow}
\title{Debug package 'futurenow'}
\usage{
debug_futurenow(tmpfile = stdout(), reset = FALSE, master_only = FALSE)
}
\arguments{
\item{tmpfile}{A temporary log file or connection; default is \code{stdout()}}

\item{reset}{Whether to reset log file if \code{tmpfile} is a local file}

\item{master_only}{Whether only to log messages from master branch}
}
\value{
Absolute path of \code{tmpfile} if \code{tmpfile} is a character,
or as-is if it is a connection.
}
\description{
Debug package 'futurenow'
}
